reorder headers in main file. rationale:
my preference is to break includes into blocks in increasing order of
standardization:

project specific includes (i.e., local things included with "") come
first, followed by a block of the 2nd least standard headers, and so
on in that fashion, with platform specific headers immediately preceding
the very last group which comprises standard C/C++ headers. ask me if it
isn't clear why this is not only a cosmetic, but also a useful practice.